Alteração do tipo decimal para intero pois estava gerando excessão no banco de dados.

+++ ContribuaMais.API/Models/Dados/Campanha.cs
@@ -8,7 +8,7 @@ namespace ContribuaMais.API.Models.Dados
         public DateTime Inicio { get; set; }
         public DateTime Fim { get; set; }
         public bool EstaAtiva { get; set; }
-        public decimal Objetivo { get; set; }
-        public decimal Arrecadado { get; set; }
+        public int Objetivo { get; set; }
+        public int Arrecadado { get; set; }
     }
 }
